fix the signature of the process method
This is an unfortunate change. It is a breaking change, obviously,
or maybe a fixing one if you want. I completely blanked on the fact
that forwarding passes the request arguments to the forward target.
So if you have an action with any :Args then the process method would
use the value of the first argument as the template name, overriding
any other decisions.